Island: Rename reporting/stolen_credentials.py to reporting/format_credentials.py
