+++ monkey/monkey_island/cc/services/reporting/__init__.py
@@ -1,0 +1,1 @@
+from .format_credentials import format_creds_for_reporting
