+++ monkey/monkey_island/cc/services/reporting/format_credentials.py
